ci: Automate integration tests
