+++ .github/workflows/integration-tests.yml
@@ -1,0 +1,33 @@
+on:
+  schedule:
+    # Run twice per day (12:45 AM & PM)
+    - cron: '45 */12 * * *'
+  push:
+    branches:
+      # Push to a branch that ends with "/trigger-int-tests" to manually trigger tests
+      - '**/trigger-int-tests'
+name: Integration Tests
+jobs:
+  all:
+    name: Integration Tests
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@master
+    - name: Cache node_modules
+      id: cache-modules
+      uses: actions/cache@v1
+      with:
+        path: node_modules
+        key: ${{ runner.OS }}-build-${{ hashFiles('package.json') }}
+    - name: npm install
+      if: steps.cache-modules.outputs.cache-hit != 'true'
+      run: npm install
+    - name: build
+      run: tsc
+    - name: integration tests
+      run: npm run test:int
+      env:
+        AZURE_SUBSCRIPTION_ID: ${{ secrets.AZURE_SUBSCRIPTION_ID}}
+        AZURE_CLIENT_ID: ${{ secrets.AZURE_CLIENT_ID}}
+        AZURE_CLIENT_SECRET: ${{ secrets.AZURE_CLIENT_SECRET}}
+        AZURE_TENANT_ID: ${{ secrets.AZURE_TENANT_ID}}
