+++ .gitignore
@@ -4,5 +4,6 @@ node_modules/
 /coverage/
 .env*
 .idea/
+tmp/
 
 .vscode/launch.json
