+++ integrationTests/configurations/node10-linux-external/serverless.yml
