+++ integrationTests/configurations/node10-linux/serverless.yml
