+++ integrationTests/configurations/node10-windows-webpack/serverless.yml
