+++ integrationTests/configurations/node10-windows/serverless.yml
