+++ integrationTests/configurations/node12-linux-external/serverless.yml
