+++ integrationTests/configurations/node12-linux/serverless.yml
