+++ integrationTests/configurations/node12-windows-webpack/serverless.yml
