+++ integrationTests/configurations/node12-windows/serverless.yml
