+++ integrationTests/configurations/python37/serverless.yml
@@ -7,6 +7,8 @@ provider:
   stage: qa
   region: West Europe
   apim: true
+  environment:
+    VARIABLE_FOO: 'foo'
 
 plugins:
   - serverless-azure-functions
