+++ integrationTests/models/configurations.ts
@@ -14,22 +14,3 @@ export interface Configurations {
   "python37": any,
   "python38": any
 }
-
-export function getConfigurationsObject(): Configurations {
-  return {
-    "node10-linux": {},
-    "node10-linux-external": {},
-    "node10-windows": {},
-    "node10-windows-webpack": {},
-    "node12-linux": {},
-    "node12-linux-external": {},
-    "node12-linux-premium": {},
-    "node12-windows": {},
-    "node12-windows-premium": {},
-    "node12-windows-webpack": {},
-    "python36": {},
-    "python36-premium": {},
-    "python37": {},
-    "python38": {}
-  }
-}
