+++ integrationTests/run.ts
@@ -1,7 +1,7 @@
-import { run, Summarizers } from "clvr"
+import { run } from "clvr"
 import { getDirectories } from "./tests/utils";
 import { offlineTest } from "./tests/offlineTest";
-import { defaultParameters } from "./parameters/default";
+import { defaultParameters } from "./models/parameters";
 import { deployTest } from "./tests/deployTest";
 
 const configArg = process.argv[2];
@@ -11,12 +11,6 @@ const configurations = (configArg)
   : getDirectories("configurations");
 
 run([
-  {
-    name: "Offline Test",
-    validations: offlineTest,
-    parameters: defaultParameters,
-    directories: configurations,
-  },
   {
     name: "Deploy Test",
     validations: deployTest,
